Feat: ResponseService and Update/Delete responses

+++ src/main/java/com/javatest/restapi/controller/CardController.java
@@ -1,9 +1,13 @@
 package com.javatest.restapi.controller;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.javatest.restapi.database.CardRepository;
 import com.javatest.restapi.model.Card;
+import com.javatest.restapi.services.ResponseService;
 
 @RestController
 @RequestMapping("/cards")
@@ -33,20 +38,36 @@ public class CardController {
     }
 
     @PostMapping
-    public void create(@RequestBody Card card){
-        repository.save(card);
+    public Card create(@RequestBody Card card){
+        return repository.save(card);
     }
 
     @PutMapping("/{id}")
-    public Card update(@PathVariable Long id, @RequestBody Card card){
+    public ResponseEntity<Map<String, Object>> update(@PathVariable Long id, @RequestBody Card card){
        
+        ResponseService responseService = new ResponseService();
+
+       if(repository.existsById(id)){
         card.setId(id);
-        return repository.save(card);
-        
+        repository.save(card);
+        return responseService.success("Cartão Atualizado com sucesso!", HttpStatus.OK);
+       }
+       else {
+        return responseService.error("Cartão não encontrado!", HttpStatus.NOT_FOUND);
+       }
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id){
-        repository.deleteById(id);
+    public ResponseEntity<Map<String, Object>> delete(@PathVariable Long id){
+
+        ResponseService responseService = new ResponseService();
+
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return responseService.success("Cartão deletado com sucesso!", HttpStatus.OK);
+        }
+        else {
+            return responseService.error("Cartão não encontrado!", HttpStatus.NOT_FOUND);
+        }
     }
 }
